% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/app-utils.R
\name{app-utils}
\alias{app-utils}
\alias{new_waiter}
\alias{waiter_html}
\alias{is_ci}
\alias{is_testing}
\alias{get_app_colors}
\alias{image_crop2}
\alias{get_comp_vec}
\alias{get_full_img}
\alias{get_replay_gif}
\alias{get_base_plot}
\alias{get_app_dir}
\alias{get_temp_dir}
\alias{create_session_dir}
\alias{clear_frames_dir}
\alias{clear_graphics_dir}
\alias{clear_temp_dir}
\alias{write_jpg}
\alias{write_png}
\alias{write_gif}
\alias{write_frame}
\alias{trigger_image_dl}
\alias{trigger_gif_dl}
\alias{trigger_vid_dl}
\alias{get_click_plot}
\title{App Utils}
\usage{
new_waiter(session = NULL)

waiter_html(msg)

is_ci()

is_testing()

get_app_colors()

image_crop2(im, xmin, xmax, ymin, ymax)

get_comp_vec(gif)

get_full_img()

get_replay_gif()

get_base_plot()

get_app_dir(...)

get_temp_dir(...)

create_session_dir()

clear_frames_dir()

clear_graphics_dir()

clear_temp_dir()

write_jpg(img)

write_png(img)

write_gif(gif, fps)

write_frame(gif, i, outpath)

trigger_image_dl(img_path, session)

trigger_gif_dl(gif_path, session)

trigger_vid_dl(gif_path, fps, session)

get_click_plot(base_plot, x, y)
}
\arguments{
\item{session}{shiny session}

\item{msg}{message for waiter screen}

\item{im}{image}

\item{xmin}{xmin}

\item{xmax}{xmax}

\item{ymin}{ymin}

\item{ymax}{ymax}

\item{gif}{gif object}

\item{...}{optional}

\item{img}{image object}

\item{fps}{frames per second}

\item{i}{frame number}

\item{outpath}{write path}

\item{img_path}{image path}

\item{gif_path}{gif path}

\item{base_plot}{base plot}

\item{x}{position of click}

\item{y}{position of click}
}
\description{
Functions required to execute and facililate an application user session.
}
\section{Functions}{
\itemize{
\item \code{new_waiter()}: create a new waiter object

\item \code{waiter_html()}: get html for waiter progress page

\item \code{is_ci()}: returns TRUE if called on CI

\item \code{is_testing()}: returns TRUE if called while testing

\item \code{get_app_colors()}: Set Plot colors

\item \code{image_crop2()}: crop image based on plot brush

\item \code{get_comp_vec()}: function to get frames of a gif only if there are changes from prior frame

\item \code{get_full_img()}: get base image for app

\item \code{get_replay_gif()}: get replay gif for app

\item \code{get_base_plot()}: get base image plot for app

\item \code{get_app_dir()}: get path to internal app files

\item \code{get_temp_dir()}: get path to session temp dir

\item \code{create_session_dir()}: create dirs on session start

\item \code{clear_frames_dir()}: clear generated frames

\item \code{clear_graphics_dir()}: clear generated graphics

\item \code{clear_temp_dir()}: clear all in temp dir

\item \code{write_jpg()}: write jpg

\item \code{write_png()}: write png

\item \code{write_gif()}: write gif

\item \code{write_frame()}: write frame in gif

\item \code{trigger_image_dl()}: trigger download

\item \code{trigger_gif_dl()}: trigger download

\item \code{trigger_vid_dl()}: trigger download

\item \code{get_click_plot()}: get plot based on click position

}}
