% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replayArt-package.R
\docType{package}
\name{replayArt-package}
\alias{replayArt}
\alias{replayArt-package}
\title{replayArt: Replay Portrait Drawing}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

A shiny app to generate a replay for a selected region of a finished portrait drawing.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/r-data-science/replayArt}
  \item Report bugs at \url{https://github.com/r-data-science/replayArt/issues}
}

}
\author{
\strong{Maintainer}: Bobby Fatemi \email{bfatemi07@gmail.com}

}
\keyword{internal}
