% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap4.R
\name{summaryBox}
\alias{summaryBox}
\title{Info / Value Box in Shiny Apps and RMarkdown}
\usage{
summaryBox(
  title = "title",
  value = "value",
  width = 4,
  icon = "fas fa-chart-bar",
  style = "primary",
  border = "left"
)
}
\arguments{
\item{title}{Text to be shown in the box}

\item{value}{Value to be shown in the box}

\item{width}{Width of Box. width = 4 means 3 boxes can be fitted (12 / 4)}

\item{icon}{Font Awesome 5 icons. E.g. "fas fa-chart-bar"}

\item{style}{Either "primary", "secondary", "info", "success", "danger", "warning"}

\item{border}{Either "left", "bottom"}
}
\description{
Info / Value Box in Shiny Apps and RMarkdown
}
