% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bootstrap4.R
\name{summaryBox2}
\alias{summaryBox2}
\title{Info / Value Box in Shiny Apps and RMarkdown}
\usage{
summaryBox2(title, value, width = 4, icon = "fas fa-chart-bar", style = "info")
}
\arguments{
\item{title}{Text to be shown in the box}

\item{value}{Value to be shown in the box}

\item{width}{Width of Box. width = 4 means 3 boxes can be fitted (12 / 4)}

\item{icon}{Font Awesome 5 icons. E.g. "fas fa-chart-bar"}

\item{style}{Either "primary", "secondary", "info", "success", "danger", "warning", "light", "dark"}
}
\description{
Info / Value Box in Shiny Apps and RMarkdown
}
