% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/write_graphic_2.R
\name{write_graphic}
\alias{write_graphic}
\title{Create and Write Plot}
\usage{
write_graphic(DT, i, outpath)
}
\arguments{
\item{DT}{Data}

\item{i}{Iteration}

\item{outpath}{Save Path}
}
\description{
Create and Write Plot
}
